+++ src/main/java/org/ahpuh/surf/post/controller/PostController.java
@@ -6,7 +6,13 @@ import org.ahpuh.surf.common.response.CursorResult;
 import org.ahpuh.surf.common.s3.S3Service;
 import org.ahpuh.surf.common.s3.S3ServiceImpl.FileStatus;
 import org.ahpuh.surf.jwt.JwtAuthentication;
-import org.ahpuh.surf.post.dto.*;
+import org.ahpuh.surf.post.dto.ExploreDto;
+import org.ahpuh.surf.post.dto.PostCountDto;
+import org.ahpuh.surf.post.dto.PostDto;
+import org.ahpuh.surf.post.dto.RecentPostDto;
+import org.ahpuh.surf.post.dto.request.PostRequestDto;
+import org.ahpuh.surf.post.dto.response.AllPostResponseDto;
+import org.ahpuh.surf.post.dto.response.PostResponseDto;
 import org.ahpuh.surf.post.service.PostService;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.MediaType;
