+++ src/main/java/org/ahpuh/surf/post/converter/PostConverter.java
@@ -4,7 +4,13 @@ import org.ahpuh.surf.category.dto.CategorySimpleDto;
 import org.ahpuh.surf.category.entity.Category;
 import org.ahpuh.surf.common.exception.EntityExceptionHandler;
 import org.ahpuh.surf.common.s3.S3ServiceImpl.FileStatus;
-import org.ahpuh.surf.post.dto.*;
+import org.ahpuh.surf.post.dto.PostDto;
+import org.ahpuh.surf.post.dto.PostScoreCategoryDto;
+import org.ahpuh.surf.post.dto.PostScoreDto;
+import org.ahpuh.surf.post.dto.RecentPostDto;
+import org.ahpuh.surf.post.dto.request.PostRequestDto;
+import org.ahpuh.surf.post.dto.response.AllPostResponseDto;
+import org.ahpuh.surf.post.dto.response.PostResponseDto;
 import org.ahpuh.surf.post.entity.Post;
 import org.ahpuh.surf.user.entity.User;
 import org.springframework.stereotype.Component;
