+++ src/main/java/org/ahpuh/surf/post/dto/response/PostResponseDto.java
@@ -1,4 +1,4 @@
-package org.ahpuh.surf.post.dto;
+package org.ahpuh.surf.post.dto.response;
 
 import lombok.*;
