+++ src/main/java/org/ahpuh/surf/post/service/PostService.java
@@ -9,6 +9,9 @@ import org.ahpuh.surf.common.response.CursorResult;
 import org.ahpuh.surf.common.s3.S3ServiceImpl.FileStatus;
 import org.ahpuh.surf.post.converter.PostConverter;
 import org.ahpuh.surf.post.dto.*;
+import org.ahpuh.surf.post.dto.request.PostRequestDto;
+import org.ahpuh.surf.post.dto.response.AllPostResponseDto;
+import org.ahpuh.surf.post.dto.response.PostResponseDto;
 import org.ahpuh.surf.post.entity.Post;
 import org.ahpuh.surf.post.repository.PostRepository;
 import org.ahpuh.surf.user.entity.User;
