+++ src/test/java/org/ahpuh/surf/unit/post/service/PostServiceTest.java
@@ -3,7 +3,7 @@ package org.ahpuh.surf.unit.post.service;
 import org.ahpuh.surf.category.entity.Category;
 import org.ahpuh.surf.category.repository.CategoryRepository;
 import org.ahpuh.surf.post.converter.PostConverter;
-import org.ahpuh.surf.post.dto.PostRequestDto;
+import org.ahpuh.surf.post.dto.request.PostRequestDto;
 import org.ahpuh.surf.post.entity.Post;
 import org.ahpuh.surf.post.repository.PostRepository;
 import org.ahpuh.surf.post.service.PostService;
